career-mentor-agent-03/pyproject.toml
[project]
name = "career-mentor-agent-03"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.1.0",
    "python-dotenv>=1.1.1",
]
